Improve behavior around tidying
When a non-branch segment was focused, tidying lost the focus (moving it
to the whole branch), and caused some other downstream misbehavior. This
adds a notion of a "root" for the graph's focus, moves the root around
when tidying, and recovers global state from the root after the tidy, in
order to make sure that the global focus is a sensible segment
corresponding to the one displayed in the graph.