Add support for multiple regexes
Multiple pruning regexes can now be entered, by seperating them with `||`
in the pruning regex input. Pruning with multiple regexes will remove
any branch that, for any of the regexes, matches that regex and also
only partners with branches that match the regex.

So entering `bad||stored` will prune all branches that have `bad` on
stdout and also only partner with branches that have `bad` on stdout,
and similarly for `stored`. This is different from `bad|stored`, which
would prune all branches that match `bad|stored` and which partner only
with branches matching this - `bad|stored` would prune a branch that had
`bad` on stdout prepatch and `stored` on stdout postpatch, for example.